chore(create) component for creating a board
